chore: move business logic to services.py
