+++ guesstheracetrack/games/urls.py
@@ -1,8 +1,8 @@
 from django.urls import path
 
 from .views import competitive_mode
+from .views import competitive_mode_quit_session
 from .views import competitive_mode_restart_session
-from .views import competitive_mode_send_segments
 from .views import famous_tracks
 from .views import famous_tracks_quit_session
 from .views import famous_tracks_restart_session
@@ -20,7 +20,7 @@ urlpatterns = [
         name="start_session",
     ),
     path(
-        "famous_tracks/session_complete/",
+        "session_complete/<str:game_type>/",
         view=session_complete,
         name="session_complete",
     ),
@@ -37,12 +37,12 @@ urlpatterns = [
     path(
         "famous_tracks/quit_session/",
         view=famous_tracks_quit_session,
-        name="quit_session",
+        name="famous_tracks_quit_session",
     ),
     path("competitive_mode/", view=competitive_mode, name="competitive_mode"),
     path(
-        "competitive_mode/start_competitive_mode/",
-        view=competitive_mode_send_segments,
-        name="start_competitive_mode",
+        "competitive_mode/quit_session/",
+        view=competitive_mode_quit_session,
+        name="competitive_mode_quit_session",
     ),
 ]
