+++ guesstheracetrack/static/css/competitive_mode.css
@@ -9,6 +9,35 @@
   border-radius: 16px;
   padding: 0;
   aspect-ratio: 250 / 150;
+  position: relative;
+}
+
+.floating-button {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  z-index: 1000;
+  background-color: #007bff;
+  color: white;
+  border: none;
+  border-radius: 50px;
+  padding: 15px 30px;
+  font-size: 16px;
+  font-weight: bold;
+  cursor: pointer;
+  box-shadow: 0 4px 15px rgba(0, 123, 255, 0.3);
+  transition: all 0.3s ease;
+}
+
+.floating-button:hover {
+  background-color: #0056b3;
+  transform: translate(-50%, -50%) scale(1.05);
+  box-shadow: 0 6px 20px rgba(0, 123, 255, 0.4);
+}
+
+.floating-button.hidden {
+  display: none;
 }
 
 .grid-row {
@@ -92,7 +121,7 @@
   margin-left: 10px;
 }
 
-.track-circle-icon{
+.track-circle-icon {
   fill: var(--bs-gray);
 }
