+++ guesstheracetrack/templates/games/famous_tracks.html
@@ -86,7 +86,8 @@
                    class="btn btn-primary btn-submit">Restart</a>
               </div>
               <div class="col-auto">
-                <a href="{% url 'games:quit_session' %}" class="btn btn-danger">Quit</a>
+                <a href="{% url 'games:famous_tracks_quit_session' %}"
+                   class="btn btn-danger">Quit</a>
               </div>
             </div>
           </div>
